feat: second half of the compact os chunks with a few negative tests
